+++ CHANGES.md
@@ -10,6 +10,8 @@ This document is intended for Spotless developers.
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+### Fixed
+* Fix license header bug for years in range ([#840](https://github.com/diffplug/spotless/pull/840)).
 
 ## [2.13.1] - 2021-04-10
 ### Changed