+++ plugin-gradle/CHANGES.md
@@ -3,6 +3,8 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `3.27.0`).
 
 ## [Unreleased]
+### Fixed
+* Fix license header bug for years in range ([#840](https://github.com/diffplug/spotless/pull/840)).
 
 ## [5.12.0] - 2021-04-10
 ### Added